audit aws EventBridge

+++ lambda/invoices/invoiceImportFunction.ts
@@ -1,5 +1,5 @@
 import { Context, S3Event, S3EventRecord } from "aws-lambda"
-import { ApiGatewayManagementApi, DynamoDB, S3 } from "aws-sdk"
+import { ApiGatewayManagementApi, DynamoDB, EventBridge, S3 } from "aws-sdk"
 import * as AWSXRay from "aws-xray-sdk"
 import { InvoiceFile, InvoiceRepository } from "/opt/nodejs/invoiceRepository"
 import { InvoiceTransactionRepository, InvoiceTransactionStatus } from "/opt/nodejs/invoiceTransaction"
@@ -9,12 +9,16 @@ AWSXRay.captureAWS(require('aws-sdk'))
 
 const invoicesDdb = process.env.INVOICE_DDB!
 const invoicesWsApiEndpoint = process.env.INVOICE_WSAPI_ENDPOINT!.substring(6)
+const auditBusName = process.env.AUDIT_BUS_NAME!
+
 
 const s3Client = new S3()
 const ddbClient = new DynamoDB.DocumentClient()
 const apigwManagementApi = new ApiGatewayManagementApi({
     endpoint: invoicesWsApiEndpoint
 })
+const eventBridgeClient = new EventBridge()
+
 
 const invoiceTransactionRepository = new InvoiceTransactionRepository(ddbClient, invoicesDdb)
 const invoiceWSService = new InvoiceWSService(apigwManagementApi)
@@ -58,37 +62,57 @@ async function processRecord(record: S3EventRecord) {
         const invoice = JSON.parse(object.Body!.toString('utf-8')) as InvoiceFile
         console.log(invoice)
 
-        const createInvoicePromise = invoiceRepository.create({
-            pk: `#invoice_${invoice.customerName}`,
-            sk: invoice.invoiceNumber,
-            ttl: 0,
-            totalValue: invoice.totalValue,
-            productId: invoice.productId,
-            quantity: invoice.quantity,
-            transactionId: key,
-            createdAt: Date.now()
-        })
+        if (invoice.invoiceNumber.length >= 5) {
+            const createInvoicePromise = invoiceRepository.create({
+                pk: `#invoice_${invoice.customerName}`,
+                sk: invoice.invoiceNumber,
+                ttl: 0,
+                totalValue: invoice.totalValue,
+                productId: invoice.productId,
+                quantity: invoice.quantity,
+                transactionId: key,
+                createdAt: Date.now()
+            })
 
-        const deleteObjectPromise = s3Client.deleteObject({
-            Key: key,
-            Bucket: record.s3.bucket.name
-        }).promise()
+            const deleteObjectPromise = s3Client.deleteObject({
+                Key: key,
+                Bucket: record.s3.bucket.name
+            }).promise()
 
-        const updateInvoicePromise = invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.PROCESSED)
+            const updateInvoicePromise = invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.PROCESSED)
 
-        const sendStatusPromise = invoiceWSService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.PROCESSED)
+            const sendStatusPromise = invoiceWSService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.PROCESSED)
 
-        await Promise.all([createInvoicePromise, deleteObjectPromise, updateInvoicePromise, sendStatusPromise])
+            await Promise.all([createInvoicePromise, deleteObjectPromise, updateInvoicePromise, sendStatusPromise])
+        } else {
+            console.error(`Invoice import failed - non valid invoice number - transactionId: ${key}`)
 
+            const putEventPromise = eventBridgeClient.putEvents({
+                Entries: [
+                    {
+                        Source: 'app.invoice',
+                        EventBusName: auditBusName,
+                        DetailType: 'invoice',
+                        Time: new Date(),
+                        Detail: JSON.stringify({
+                            errorDetail: 'FAIL_NO_INVOICE_NUMBER',
+                            info: {
+                                invoiceKey: key,
+                                customerName: invoice.customerName
+                            }
+                        })
+
+                    }
+                ]
+            }).promise()
+
+            const sendStatusPromise = invoiceWSService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.NON_VALID_INVOICE_NUMBER)
+
+            const updateInvoicePromise = invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.NON_VALID_INVOICE_NUMBER)
+
+            await Promise.all([sendStatusPromise, updateInvoicePromise, putEventPromise])
+        }
     } catch (error) {
-        const invoiceTransaction = await invoiceTransactionRepository.getInvoiceTransaction(key)
-
-        console.error(`Invoice import failed - non valid invoice number - transactionId: ${key}`)
-
-        const sendStatusPromise = invoiceWSService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.NON_VALID_INVOICE_NUMBER)
-
-        const updateInvoicePromise = invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.NON_VALID_INVOICE_NUMBER)
-
-        await Promise.all([sendStatusPromise, updateInvoicePromise])
+        console.log((<Error>error).message)
     }
 }
